move Operator sturct to middlewate
